* libz-sys now allows to build from cc instead of cmake. This patch enables this
  for the flate2 crate, which should be the fastest implementation of gzip
* Update LAST_RELEASE to v0.18.0
* Use libz-sys instead of gzip-zune-inflate when building release binaries
```
Old(gzip-zune-inflate):
Benchmark 2: backhand-dist
  Time (mean ± σ):     804.6 ms ±  12.8 ms    [User: 1047.7 ms, System: 267.5 ms]
  Range (min … max):   771.8 ms … 837.3 ms    50 runs

Benchmark 3: backhand-dist-musl
  Time (mean ± σ):      1.399 s ±  0.019 s    [User: 1.903 s, System: 0.330 s]
  Range (min … max):    1.358 s …  1.442 s    50 runs

New(flate::libz-sys):
Benchmark 2: backhand-dist
  Time (mean ± σ):     672.0 ms ±   5.9 ms    [User: 846.2 ms, System: 217.1 ms]
  Range (min … max):   659.2 ms … 693.9 ms    50 runs

Benchmark 3: backhand-dist-musl
  Time (mean ± σ):     786.2 ms ±   5.6 ms    [User: 1006.3 ms, System: 203.1 ms]
  Range (min … max):   776.5 ms … 808.8 ms    50 runs
```

Closes #556